Fix crash when switching tab on minigame info
Fixes #66

This was caused by the new model getting the wrong name. The reason it
got the wrong name was because it copied it from the existing player
node and then the player node called queue_free() but this does not
delete the player node before the new node is added which makes it the
new model take a different name.

Signed-off-by: Jakob Sinclair <sinclair.jakob@mailbox.org>