Improve outlines
Replaced the outline shader with the outline shader by Nathan Lovato (GDQuest), because it gives better results, when there are not fully transparent pixels around the edges of the image.

Also added a bit of space to the (green) tux and the cookie trap icons, so that the outline is not cut off when the image ends.

And a lot of cleanup related to that.